pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axiomatic-mcp"
version = "1.0.0"
description = "Modular MCP servers for Axiomatic"
authors = [{name = "Axiomatic Team", email = "team@axiomatic.ai"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
axiomatic-pic = "axiomatic_mcp.domains.pic:main"
axiomatic-electronics = "axiomatic_mcp.domains.electronics:main"
axiomatic-quantum = "axiomatic_mcp.domains.quantum:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["axiomatic_mcp*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "UP", "B", "C4", "PTH", "SIM", "RUF"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true